+++ app/src/main/java/com/untilled/roadcapture/data/datasource/api/dto/user/FollowingsCondition.kt
@@ -1,6 +1,6 @@
 package com.untilled.roadcapture.data.datasource.api.dto.user
 
 data class FollowingsCondition(
-    var userId: Int,
+    var userId: Long,
     var username: String? = null
 )
