+++ app/src/main/java/com/untilled/roadcapture/data/datasource/dao/LocalUserDao.kt
@@ -3,7 +3,7 @@ package com.untilled.roadcapture.data.datasource.dao
 import com.untilled.roadcapture.data.datasource.api.dto.user.UserResponse
 
 interface LocalUserDao {
-    fun saveUser(id: Int)
-    fun getUser(): Int
+    fun saveUser(id: Long)
+    fun getUser(): Long
     fun clearUser()
 }
